+++ src/main/java/com/donkeys_today/server/support/jwt/JWTUtil.java
@@ -1,0 +1,68 @@
+package com.donkeys_today.server.support.jwt;
+import static com.donkeys_today.server.support.dto.type.ErrorType.EXPIRED_TOKEN;
+import static com.donkeys_today.server.support.dto.type.ErrorType.INVALID_TOKEN;
+import static com.donkeys_today.server.support.dto.type.ErrorType.UNKNOWN_TOKEN;
+import static com.donkeys_today.server.support.dto.type.ErrorType.UNSUPPORTED_TOKEN;
+import static com.donkeys_today.server.support.dto.type.ErrorType.WRONG_SIGNATURE_TOKEN;
+
+import com.donkeys_today.server.support.exception.UnauthorizedException;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.MalformedJwtException;
+import io.jsonwebtoken.UnsupportedJwtException;
+import io.jsonwebtoken.security.SignatureException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class JWTUtil {
+
+    private final JwtTokenProvider jwtTokenProvider;
+
+    public JwtValidationType validateToken(String token) {
+        try {
+            final Claims claims = getBody(token);
+            if (claims.get(JWTConstants.TOKEN_TYPE).toString().equals(JWTConstants.ACCESS_TOKEN)) {
+                return JwtValidationType.VALID_ACCESS;
+            } else if (claims.get(JWTConstants.TOKEN_TYPE).toString().equals(JWTConstants.REFRESH_TOKEN)) {
+                return JwtValidationType.VALID_REFRESH;
+            }
+            throw new UnauthorizedException(INVALID_TOKEN);
+        } catch (MalformedJwtException e) {
+            throw new UnauthorizedException(INVALID_TOKEN);
+        } catch (ExpiredJwtException e) {
+            throw new UnauthorizedException(EXPIRED_TOKEN);
+        } catch (IllegalArgumentException e) {
+            throw new UnauthorizedException(UNKNOWN_TOKEN);
+        } catch (UnsupportedJwtException e) {
+            throw new UnauthorizedException(UNSUPPORTED_TOKEN);
+        } catch (SignatureException e) {
+            throw new UnauthorizedException(WRONG_SIGNATURE_TOKEN);
+        }
+    }
+
+    private Claims getBody(final String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(jwtTokenProvider.getSigningKey())
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    public String getUserFromJwt(String token) {
+        Claims claims = getBody(token);
+        return claims.get(JWTConstants.USER_ID).toString();
+    }
+
+    public String getRoleFromJwt(String token) {
+        Claims claims = getBody(token);
+        return claims.get(JWTConstants.ROLE).toString();
+    }
+
+    public String getTypeFromJwt(String token) {
+        Claims claims = getBody(token);
+        return claims.get(JWTConstants.TOKEN_TYPE).toString();
+    }
+}
