+++ src/test/java/com/donkeys_today/server/docs/JwtTokenProviderTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import com.donkeys_today.server.support.exception.UnauthorizedException;
 import com.donkeys_today.server.support.jwt.JWTConstants;
+import com.donkeys_today.server.support.jwt.JWTUtil;
 import com.donkeys_today.server.support.jwt.JwtTokenProvider;
 import com.donkeys_today.server.support.jwt.JwtValidationType;
 import org.junit.jupiter.api.DisplayName;
@@ -16,6 +17,9 @@ public class JwtTokenProviderTest {
 
     @Autowired
     private JwtTokenProvider jwtTokenProvider;
+
+    @Autowired
+    private JWTUtil jwtUtil;
 
     @Test
     @DisplayName("액세스 토큰 발급 테스트")
@@ -49,7 +53,7 @@ public class JwtTokenProviderTest {
     @DisplayName("JWT에서 사용자 ID 가져오기 테스트")
     public void testGetUserFromJwt() {
         String token = jwtTokenProvider.issueAccessToken(1L, "USER");
-        String userId = jwtTokenProvider.getUserFromJwt(token);
+        String userId = jwtUtil.getUserFromJwt(token);
         assertEquals("1", userId);
     }
 
@@ -57,7 +61,7 @@ public class JwtTokenProviderTest {
     @DisplayName("JWT에서 역할 가져오기 테스트")
     public void testGetRoleFromJwt() {
         String token = jwtTokenProvider.issueAccessToken(1L, "USER");
-        String role = jwtTokenProvider.getRoleFromJwt(token);
+        String role = jwtUtil.getRoleFromJwt(token);
         assertEquals("USER", role);
     }
 
@@ -65,7 +69,7 @@ public class JwtTokenProviderTest {
     @DisplayName("JWT에서 유형 가져오기 테스트")
     public void testGetTypeFromJwt() {
         String token = jwtTokenProvider.issueAccessToken(1L, "USER");
-        String type = jwtTokenProvider.getTypeFromJwt(token);
+        String type = jwtUtil.getTypeFromJwt(token);
         assertEquals(JWTConstants.ACCESS_TOKEN, type);
     }
 
@@ -73,13 +77,13 @@ public class JwtTokenProviderTest {
     @DisplayName("토큰 유효성 검사 - 유효한 액세스 토큰 테스트")
     public void testValidateToken_ValidAccessToken() {
         String token = jwtTokenProvider.issueAccessToken(1L, "USER");
-        JwtValidationType validationType = jwtTokenProvider.validateToken(token);
+        JwtValidationType validationType = jwtUtil.validateToken(token);
         assertEquals(JwtValidationType.VALID_ACCESS, validationType);
     }
 
     @Test
     @DisplayName("토큰 유효성 검사 - 잘못된 토큰 테스트")
     public void testValidateToken_InvalidToken() {
-        assertThrows(UnauthorizedException.class, () -> jwtTokenProvider.validateToken("invalidtoken"));
+        assertThrows(UnauthorizedException.class, () -> jwtUtil.validateToken("invalidtoken"));
     }
 }
